requirements.txt
# MCP Server Dependencies
# Python 3.12.12 compatible

# Core MCP SDK
mcp>=1.0.0
fastmcp>=0.1.0

# HTTP & API clients
httpx>=0.27.0
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Caching (Redis)
redis>=5.0.0

# Anthropic Claude API
anthropic>=0.25.0

# ASGI server (for production)
uvicorn[standard]>=0.27.0

# Data validation
pydantic>=2.6.0

# Utilities
python-dateutil>=2.8.2
